+++ app/layouts/default.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="max-w-480 min-h-screen mx-auto flex flex-col justify-start items-center gap-5 bg-[#F6F6F6]">
     <LayoutNavbarUser />
-    <div class="px-20 md:px-36 lg:px-64 mt-5">
+    <div class="px-20 md:px-36 lg:px-80 mt-5 w-full">
       <slot />
     </div>
   </div>